added files to gitignore
I checked out telegram and added files to be excluded from git
to .gitignore .

created 2 utilities classes, one is used to wake your phone up
when it receives a GCM message.
the other one handles the server side aspects for the application.

The GCMintentservice defines how the app will respond to the
various occurrences related to an incoming GCMmessage.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>