Implemented the article list
I have added the list of custom views which shows the articles
along with an image.
The app still crashes if we rotate it too fast more than twice.
Removing the progress bar stops that bug.

So the app mainly has GCm and encoding issue and catogeries left to
work on.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>