Loads of changes
So, I realized that GCM was a library GCMRegistrar which now
is a API namely GoogleCloudMessaging.
So, as vishnu got the library up and running, I am getting the API to
work in this app, I have added the google-services.json file which I
got from the developers console.

The Google Services plugin for Gradle parses configuration
information from the google-services.json file.
I added the plugin as a dependency to the app module and the top
level gradle file.

I got the most awesome gitignore, its explained in the file itself all
the files which are being ignored

Now I got the GCM-quickstart up and running and got a good Idea of
how it works there, I am working from that and the guide on the
developers site to get GCM up on this one, without using deprecated
libraries.

I have Set Up Google Play Services, edited the manifest to allow all
required permission.

I have got the check for Google Play Services APK in place.

Now before moving on to writing code to Obtain a registration token,

I thought hey this is some serious stuff I am writing, better commit it.

So, the entire Utilities pakage has been looked through.
In UI GCMData and Splash screen have been refactored.

sorry for the large commit.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>