Polished SplashScreen
I got a progressBar even on the splash screen, It now checks for the
internet connection in the background While displaying the splashscreen
image. (before it used to show for a brief amount of timr a white screen
, now the transition from clicking on the app to the app opening is
smooth)

I also kept the checking for internet AND intent launch to the next
activity in the onStart so that the user can go switch on his internet
and return to the app. Before he would have to repen the app.

Now thinking back onCreate was a really stupid place to keep that code.

Well, hopefully I am learning from my mistakes.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>