Used custom fragments and RecyclerView
So, I checked out the sample app they had kept for navigationDrawer
and understood how I was supposed to display the different categories.

To have a better UI, I used GridLayout when the orientation is landscape

The app still has problems due to the fact that the fragments are being
updated asyncronously.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>