Fixed UI
I got the gridview layout such that a single column in portrait
and 2 columns in the landscape mode.

I got the banners working for the display article.

I have both orientations working in this app, unlike other news apps.

I got support for the HTML tags sent over by the server.

I used a black and white color combination UI.

I have made it a endless adapter for the scrolling thus have lazy
loading up, the transition is not smooth.

The app need to have the settings implemented.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>