Added Masthead, Icon, improved net connectivity checker
I got the App Icon resource from the company and added that as
the masthead for the app.

I had the connection checker to try to ping the motorindia server 10 times
because on un reliable networks packet loss is expected.

I added settings but I have yet to have the changed sharedpreference value
actually affect the notifications yet.

Signed-off-by: Aditya A Prasad <adityaprasadiskool@gmail.com>