Release V1.0
This has been released to the Play Store
[here](https://play.google.com/store/apps/details?id=in.motorindiaonline.motorindia)

The main client side errors as of V1 are

On pressing back from the display article activity (where you view the article)
you arent send back to the category you were browsing before that. It goes to the featured
article.

on choosing a category there is a significant dely before the article list loads.
and the article list loading is asynchronous therefore if you choose a category
then choose cataloges quickly, the cataloges image is overridden by the choosen categorie's
fragment.

The server side of GCM is yet to be fixed, it seems like it should work as soon as the server side
is fixed, but the length of the UID i am using might cause problems.

On the whole, this really simple sounding application has taken a lot of work to complete
but I cannot honestly say I am proud.

Signed-off-by: Aditya A Prasad <adityaprasadiskool@gmail.com>